+++ tests/v2/guest_user.spec.js
@@ -18,7 +18,9 @@ test('Guest user', async ({ page }) => {
     await test.step("Make the user a guest", async () => {
         await page.goto(`/v2/admin/users/${userId}/edit`);
         await waitPageLoading(page);
+        await expect(page.getByRole('checkbox', { name: 'Superuser' })).toBeEnabled();
         await page.getByRole('checkbox', { name: 'Guest' }).check();
+        await expect(page.getByRole('checkbox', { name: 'Superuser' })).not.toBeEnabled();
         await page.getByRole('button', { name: 'Save' }).click();
         await expect(page.getByText('User successfully updated')).toBeVisible();
     })
@@ -28,19 +30,5 @@ test('Guest user', async ({ page }) => {
         await waitPageLoading(page);
         const userRow = await page.getByRole('row', { 'name': userEmail }).getByRole('cell').all();
         await verifyChecked(userRow, 5, true);
-    })
-
-    await test.step("Verify superuser cannot be guest", async () => {
-        await page.goto(`/v2/admin/users/${userId}/edit`);
-        await waitPageLoading(page);
-        await page.getByRole('checkbox', { name: 'Superuser' }).check();
-        await page.getByRole('button', { name: 'Save' }).click();
-        await page.getByRole('button', { name: 'Confirm' }).click();
-        await expect(page.getByText('Superuser cannot be guest.')).toBeVisible();
-
-        await page.getByRole('checkbox', { name: 'Guest' }).uncheck();
-        await page.getByRole('button', { name: 'Save' }).click();
-        await page.getByRole('button', { name: 'Confirm' }).click();;
-        await expect(page.getByText('User successfully updated')).toBeVisible();
     })
 })